requirements.txt
Django==5.1
requests
python-dotenv
djangorestframework
transformers